Rather than a single build_target being shared between multiple test_runs, rework the data mdoel so that each test_run has a single build_target (it was how the importer worked anyway). This required reordering of the migrations to ensure correct order of foreign key creation.
Rework the routes so that build_configuration and build_target are both below test_run in the URI namespace. This makes sense now that they are specific to a particular test_run. It also means that our cache clearing protocol will clear these directories.

Update the views to use link_for() to link to these and other objects so correct urls are generated.