Refactor code so that importing test data is no longer done through a web interface but is done by placing the results into designated directories on the server and running a separate importer process.
The reason for this is because ant has no easy way to POST multipart data to a web server but we can copy it to a specific directory. It also means that the mongrel will never need load balancing due to large uploads occuring.

This change required removing the concept of uploader throughout the system and the enhancement and extension of the existing import code. It also involved removal of web interface for upload.