Extend test-case by adding optional test-compilation element.
A test-case may have zero or one test-compilation children, which
capture the output from compiling the test case.  This isn't particularly
useful for Jikes RVM, where compiling the test case isn't an integral part of
the test.  It is quite useful for X10, where the compiler being used to compile
the test case is part of what is being tested.