Initial prototype of a page to compare two arbitrary test runs and catagorize test cases within the two runs by comparing their results.
Still some minor UI improvements planned, but this is a good point 
for an initial commit of the functionality.

Minor refactoring of test_link helper (to take test_run as a param)
also included in this commit.  Necessary to allow helper to be used
to generate links to the details of test cases from either of the runs
being compared.