Merge of the upgrade-to-rails-2.2 branch to the trunk.
The changes don't actually get us to working on rails 2.2, 
but do include a number of changes that make it easier to 
do forward development of cattrack and deploy it to a mongrel
cluster served by a proxy-balancing httpd via capistrano.

Will re-visit upgrading rails version in the future, 
most likely on the trunk now that the machinery is 
in place to more easily roll deployed versions 
forward/backwards.